+++ recognition/Emitter.py
@@ -1,0 +1,8 @@
+from pymitter import EventEmitter
+
+event = EventEmitter()
+
+import Actions
+import MultiGesture
+
+# event.emit("start", hand="test", gest="test")
